+++ src/bsp/include/bsp/IBSP.h
@@ -6,7 +6,7 @@ class IBSP {
     virtual ~IBSP() = default;
 
     /**
-     * @brief Initialise the
+     * @brief Initialise the chip for usage. Needs to be called early in the program.
      */
     virtual void initChip() = 0;
 };
