+++ src/bsp/src/stm32f4/hal/include/hal.h
@@ -1,0 +1,5 @@
+/**
+ * @brief  Retargets the C library printf function to the USART.
+ * @param  None
+ * @retval None
+ */
