+++ src/bsp/src/stm32f4/hal/include/uart_print.h
@@ -1,0 +1,4 @@
+/**
+ * @brief  Initialize the uart print, after this call, you can use printf as usual
+ */
+void UartPrint_init();
