+++ src/bsp/src/stm32f4/hal/src/hal.c
@@ -1,0 +1,13 @@
+#include "uart_print.h"
+void Hal_init(){
+    UartPrint_init();
+    /* Reset of all peripherals, Initializes the Flash interface and the Systick. */
+    HAL_Init();
+
+    /* Configure the system clock */
+    SystemClock_Config();
+
+    /* Initialize all configured peripherals */
+    MX_GPIO_Init();
+    MX_USART3_UART_Init();
+}
