+++ src/bsp/src/stm32f4/hal/stm32f429zi/Core/Inc/hivemind_hal.h
@@ -10,6 +10,10 @@ extern "C" {
 #include "stm32f4xx_hal_conf.h"
 #include "usart.h"
 
+
+#define HUART_PRINT (&huart3)
+
+
 void SystemClock_Config(void);
 
 #ifdef __cplusplus
