+++ src/bsp/src/stm32f4/hal/stm32f429zi/Core/Src/main.c
@@ -24,8 +24,7 @@
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
-#include "hivemind_hal.h"
-#include <stm32f4xx_hal_uart.h>
+
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -35,13 +34,7 @@
 
 /* Private define ------------------------------------------------------------*/
 /* USER CODE BEGIN PD */
-#ifdef __GNUC__
-/* With GCC, small printf (option LD Linker->Libraries->Small printf
-   set to 'Yes') calls __io_putchar() */
-#define PUTCHAR_PROTOTYPE int __io_putchar(int ch)
-#else
-#define PUTCHAR_PROTOTYPE int fputc(int ch, FILE* f)
-#endif /* __GNUC__ */
+
 /* USER CODE END PD */
 
 /* Private macro -------------------------------------------------------------*/
@@ -63,31 +56,6 @@ void SystemClock_Config(void);
 
 /* Private user code ---------------------------------------------------------*/
 /* USER CODE BEGIN 0 */
-
-/**
- * @brief  Retargets the C library printf function to the USART.
- * @param  None
- * @retval None
- */
-PUTCHAR_PROTOTYPE {
-    /* Place your implementation of fputc here */
-    /* e.g. write a character to the USART3 and Loop until the end of transmission */
-    volatile CircularBuffRet ret;
-    if (HAL_UART_GetState(&huart3) == HAL_UART_STATE_READY) {
-        HAL_UART_Transmit_IT(&huart3, (uint8_t*)&ch, 1);
-    } else {
-        ret = CircularBuff_putc(&cbuffUart3, ch);
-        // Error handling
-        if (ret != CircularBuff_Ret_Ok) {
-            uint8_t buffErrMsg[] = "UART3 buffer full, clearing\r\n";
-            lastUart3TransferSize = 0;
-            HAL_UART_Abort(&huart3);
-            HAL_UART_Transmit(&huart3, buffErrMsg, sizeof(buffErrMsg), HAL_MAX_DELAY);
-            CircularBuff_clear(&cbuffUart3);
-        }
-    }
-    return ch;
-}
 
 /* USER CODE END 0 */
